+++ charge-pile-machine-server/src/main/java/com/huamar/charge/pile/entity/dto/resp/McAuthResp.java
@@ -48,6 +48,6 @@ public class McAuthResp extends BaseResp {
 	/**
 	 * 原消息序号
 	 */
-	private int msgNumber;
+	private short msgNumber;
 
 }
