+++ charge-pile-machine-server/src/main/java/com/huamar/charge/pile/server/service/parameter/PileParameterExecute.java
@@ -64,7 +64,7 @@ public interface PileParameterExecute<T extends PileBaseParameterDTO> {
 
         DataPacketWriter writer = this.writer(command);
         byte[] bytes = writer.toByteArray();
-        packet.setMsgBodyLen(bytes.length);
+        packet.setMsgBodyLen((short) bytes.length);
         packet.setMsgBody(bytes);
         return packet;
     }
